// CSKT/originaldb-gate-fe/src/modules/operation-categories/types/MicrowaveLines.types.ts
export type MicrowaveLinesData = {
  antenna_height?: number | null;
  bit_rate?: number | null;
  capacity?: number | null;
  date_created?: string | null;
  date_updated?: string | null;
  depreciation_rate?: number | null;
  distance?: number | null;
  id: string;
  is_enable?: boolean | null;
  latitude?: string | null;
  longitude?: string | null;
  modulation_type?: string | null;
  name?: string | null;
  note?: string | null;
  number_of_relay_stages?: number | null;
  order_number?: number | null;
  org_manage_code_old?: string | null;
  org_manage_id?: string | null;
  polarization_type?: string | null;
  receiving_frequency?: number | null;
  receiving_sensitivity?: number | null;
  symbol?: string | null;
  transmission_frequency?: number | null;
  transmission_power?: number | null;
  type?: string | null;
};


// Công suất phát  transmittal power
// Số chặng tiếp phát number_of_relay_stages
// Tôc độ bít (R bít) bit_rate

// Dung lượng hệ thống capacity

// Độ cao anten  antenna_height
// Phân cực 
// Mức suy hao


// transmittal power


// number_of_relay_stages


// bit_rate


// capacity


// antenna_height